build mac and linux prebuilts
